added the function to a seperate utils file
