+++ src/components/Modal/Content/UpdateNotice.jsx
@@ -10,6 +10,7 @@ import { Notice } from '@wordpress/components';
 import { addQueryArgs } from '@wordpress/url';
 import { createInterpolateElement } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
+import { formatVersion } from '../../../helpers';
 
 /**
  * Internal dependencies
@@ -21,28 +22,6 @@ import {
 } from '../../../constants';
 
 const UpdateNotice = () => {
-	/* To format the version to have semver MAJOR.MINOR.PATCH. Adding '0', if the MINOR or PATCH are missing  */
-	function formatVersion(version) {
-		const hasMinorAndPatch = /^\d+\.\d+\.\d+/.test(version);
-
-		if (hasMinorAndPatch) {
-			return version;
-		}
-		/* For a version that looks like 1.2.3-RC1, numericVersion = "1.2.3" rcSuffix = "RC1" */
-		const [numericVersion, rcSuffix] = version.split(/-(.+)/);
-		const versionParts = numericVersion.split('.');
-
-		while (versionParts.length < 3) {
-			versionParts.push('0');
-		}
-
-		const formattedVersion = rcSuffix
-			? `${versionParts.join('.')}-${rcSuffix}`
-			: versionParts.join('.');
-
-		return formattedVersion;
-	}
-
 	try {
 		if (
 			compare(
