+++ src/helpers/index.js
@@ -2,3 +2,4 @@ export * from './analytics';
 export * from './blockInserter';
 export * from './fetcher';
 export * from './optimizePreview';
+export * from './utils';
